restore usage `/medicine/[slug]` page for all `medicine` pages
